ulong

+++ DbUtils/TableColumn.cs
@@ -13,7 +13,7 @@ namespace DbUtils
         public string type { get; set; }
         public bool null_able { get; set; }
         public string comments { get; set; }
-        public int? character_maximum_length { get; set; }
+        public ulong? character_maximum_length { get; set; }
         public int? numeric_precision { get; set; }
         public int? numeric_scale { get; set; }
     }
